requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# FastAPI & Web
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2

# Database
supabase==2.6.0

# LLM Providers
anthropic>=0.25.0
openai>=1.30.0

